feat: Add variants of if and define
`if` and `define` now each take variants with one less argument.

Users can now omit the "else" clause of the `if` expression. If the
condition is false, nothing happens and `()` is returned.

Users can omit the "value" clause of the `define` expression. The value
of the expression in the environment is set to `true`.